pyproject.toml
[project]
name = "veridian"
version = "0.1.0"
description = "Veridian Misinformation Response Engine — AI-native multimodal misinformation detection platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # ── API Framework ────────────────────────────────────────────────
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic[email]>=2.0",
    "pydantic-settings>=2.0",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "python-multipart>=0.0.9",

    # ── Structured Logging ───────────────────────────────────────────
    "structlog>=24.0",

    # ── Database ─────────────────────────────────────────────────────
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "redis>=5.0",
    "qdrant-client>=1.9",
    "neo4j>=5.19",

    # ── Task Queue ───────────────────────────────────────────────────
    "celery[redis]>=5.3",

    # ── ML / AI ──────────────────────────────────────────────────────
    "torch>=2.2",
    "torchvision>=0.17",
    "transformers>=4.40",
    "huggingface-hub>=0.23",
    "opencv-python-headless>=4.9",
    "numpy>=1.26",
    "pillow>=10.3",

    # ── LLM / Agents ────────────────────────────────────────────────
    "langgraph>=0.2",
    "langchain-groq>=0.1",
    "tavily-python>=0.3",

    # ── Bot ──────────────────────────────────────────────────────────
    "python-telegram-bot>=21.0",
    "httpx>=0.27",

    # ── Object Storage ───────────────────────────────────────────────
    "minio>=7.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "mypy>=1.10",
    "ruff>=0.4",
]

audio = [
    "resemblyzer>=0.1",
    "openai-whisper>=20240930",
]

gpu = [
    "facenet-pytorch>=2.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend", "ml", "workers", "bot", "scripts"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
strict = true
exclude = ["venv", ".venv", "dashboard", "extension"]
plugins = ["pydantic.mypy"]

[tool.ruff]
line-length = 120
target-version = "py311"
